chores: Reorganise files and folder
Created both entity & pojo folder and shift files accordingly
