+++ crypto-service/src/main/java/com/personal/crypto/system/crypto_service/repository/AggregatedPriceRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.personal.crypto.system.crypto_service.model.AggregatedPrice;
+import com.personal.crypto.system.crypto_service.model.entity.AggregatedPrice;
 
 public interface AggregatedPriceRepository extends JpaRepository<AggregatedPrice, Long> {
     // Result might not exists, hence optional
